refactor(bindings): move rewards types to separate module
