+++ packages/bindings/src/lib.rs
@@ -1,13 +1,17 @@
 mod msg;
 mod pagination;
 mod query;
+mod types;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub mod testing;
 
-pub use msg::{ArchwayMsg, WithdrawRewardsResponse};
+pub use msg::ArchwayMsg;
 pub use pagination::{PageRequest, PageResponse};
-pub use query::{ArchwayQuery, ContractMetadataResponse, RewardsRecord, RewardsRecordsResponse};
+pub use query::ArchwayQuery;
+pub use types::rewards::{
+    ContractMetadataResponse, RewardsRecord, RewardsRecordsResponse, WithdrawRewardsResponse,
+};
 
 pub type Coins = Vec<cosmwasm_std::Coin>;
