+++ packages/bindings/src/query.rs
@@ -2,7 +2,7 @@ use cosmwasm_std::CustomQuery;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::{pagination::PageRequest, Coins, PageResponse};
+use crate::pagination::PageRequest;
 
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -42,28 +42,4 @@ impl ArchwayQuery {
             pagination: Some(pagination),
         }
     }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct ContractMetadataResponse {
-    pub owner_address: String,
-    pub rewards_address: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct RewardsRecordsResponse {
-    pub records: Vec<RewardsRecord>,
-    pub pagination: Option<PageResponse>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct RewardsRecord {
-    pub id: u64,
-    pub rewards_address: String,
-    pub rewards: Coins,
-    pub calculated_height: i64,
-    pub calculated_time: String,
 }
